+++ src/hdmf/backends/hdf5/h5tools.py
@@ -159,25 +159,34 @@ class HDF5IO(HDMFIO):
         return file_obj
 
     @classmethod
-    @docval({'name': 'namespace_catalog', 'type': (NamespaceCatalog, TypeMap),
-             'doc': 'the NamespaceCatalog or TypeMap to load namespaces into'},
-            {'name': 'path', 'type': (str, Path), 'doc': 'the path to the HDF5 file', 'default': None},
-            {'name': 'namespaces', 'type': list, 'doc': 'the namespaces to load', 'default': None},
-            {'name': 'file', 'type': File, 'doc': 'a pre-existing h5py.File object', 'default': None},
-            {'name': 'driver', 'type': str, 'doc': 'driver for h5py to use when opening HDF5 file', 'default': None},
-            {'name': 'aws_region', 'type': str, 'doc': 'If driver is ros3, then specify the aws region of the url.',
-             'default': None},
-            returns=("dict mapping the names of the loaded namespaces to a dict mapping included namespace names and "
-                     "the included data types"),
-            rtype=dict)
+    @docval(
+        {
+            'name': 'namespace_catalog',
+            'type': (NamespaceCatalog, TypeMap),
+            'doc': 'the NamespaceCatalog or TypeMap to load namespaces into'
+        },
+        {'name': 'path', 'type': (str, Path), 'doc': 'the path to the HDF5 file', 'default': None},
+        {'name': 'namespaces', 'type': list, 'doc': 'the namespaces to load', 'default': None},
+        {'name': 'file', 'type': File, 'doc': 'a pre-existing h5py.File object', 'default': None},
+        {'name': 'driver', 'type': str, 'doc': 'driver for h5py to use when opening HDF5 file', 'default': None},
+        {
+            'name': 'aws_region',
+            'type': str,
+            'doc': 'If driver is ros3, then specify the aws region of the url.',
+            'default': None
+        },
+        returns=("dict mapping the names of the loaded namespaces to a dict mapping included namespace names and "
+                    "the included data types"),
+        rtype=dict
+    )
     def load_namespaces(cls, **kwargs):
-        """Load cached namespaces from a file.
+        """Load cached namespaces from a file into the provided NamespaceCatalog or TypeMap.
 
         If `file` is not supplied, then an :py:class:`h5py.File` object will be opened for the given `path`, the
         namespaces will be read, and the File object will be closed. If `file` is supplied, then
         the given File object will be read from and not closed.
 
-        :raises ValueError: if both `path` and `file` are supplied but `path` is not the same as the path of `file`.
+        :raises ValueError: if both `path` and `file` are supplied but `path` is not the same as the path of `file`
         """
         namespace_catalog, path, namespaces, file_obj, driver, aws_region = popargs(
             'namespace_catalog', 'path', 'namespaces', 'file', 'driver', 'aws_region', kwargs)
@@ -188,12 +197,25 @@ class HDF5IO(HDMFIO):
                 return cls.__load_namespaces(namespace_catalog, namespaces, open_file_obj)
         return cls.__load_namespaces(namespace_catalog, namespaces, open_file_obj)
 
+    @docval(
+        {
+            'name': 'namespace_catalog',
+            'type': (NamespaceCatalog, TypeMap),
+            'doc': 'the NamespaceCatalog or TypeMap to load namespaces into'
+        },
+        {'name': 'namespaces', 'type': list, 'doc': 'the namespaces to load', 'default': None}
+    )
+    def load_namespaces_io(self, **kwargs):
+        """Load cached namespaces from this HDF5IO object into the provided NamespaceCatalog or TypeMap."""
+        namespace_catalog, namespaces = getargs('namespace_catalog', 'namespaces', kwargs)
+        if not self.__file:
+            raise UnsupportedOperation("Cannot load namespaces from closed HDF5 file '%s'" % self.source)
+        return self.__load_namespaces(namespace_catalog, namespaces, self.__file)
+
     @classmethod
     def __load_namespaces(cls, namespace_catalog, namespaces, file_obj):
-        d = {}
-
         if not cls.__check_specloc(file_obj):
-            return d
+            return {}
 
         namespace_versions = cls.__get_namespaces(file_obj)
 
@@ -205,11 +227,9 @@ class HDF5IO(HDMFIO):
         for ns in namespaces:
             latest_version = namespace_versions[ns]
             ns_group = spec_group[ns][latest_version]
-            reader = H5SpecReader(ns_group)
-            readers[ns] = reader
+            readers[ns] = H5SpecReader(ns_group)
 
-        d.update(namespace_catalog.load_namespaces(cls.__ns_spec_path, reader=readers))
-
+        d = namespace_catalog.load_namespaces(cls.__ns_spec_path, reader=readers)
         return d
 
     @classmethod